Convert instruction system from long to int for better performance
Core Changes:
- Update Instruction.Value from long to int type
- Update Instruction constructor to accept int instead of long
- Change VirtualMachine stack from Stack<long> to Stack<int>
- Update VirtualMachine.Execute return type from long to int
- Change Parser.ParseLoad to use int.TryParse instead of long.TryParse
- Update JitCompiler.CompiledFunction delegate to return int instead of long

JIT Compiler Updates:
- Update JitCompilerArm64.EmitLoad to accept int parameter
- Update JitCompilerX64.EmitLoad to accept int parameter
- Optimize x64 EmitLoad to use 32-bit immediate instruction (mov r32, imm32)

Test Updates:
- Update SampleFileIntegrationTests parameterized test to use int expectedResult
- Update ErrorHandlingTests to use int.MaxValue/int.MinValue instead of long
- All existing test values remain valid within int range

Benefits:
- Improved performance with 32-bit operations instead of 64-bit
- Reduced memory usage for instruction values
- More appropriate data type for typical JIT compiler scenarios
- Maintains full compatibility with existing sample files and test cases
- All 118 tests continue to pass with 100% success rate

+++ SimpleJIT.Core/Instruction.cs
@@ -14,9 +14,9 @@ public enum InstructionType
 public class Instruction
 {
     public InstructionType Type { get; set; }
-    public long Value { get; set; }
+    public int Value { get; set; }
 
-    public Instruction(InstructionType type, long value = 0)
+    public Instruction(InstructionType type, int value = 0)
     {
         Type = type;
         Value = value;
